credit_dashboard/pyproject.toml
[project]
name = "credit_dashboard"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "camilo_bpt", email = ""},
]
dependencies = ["pyspark>=3.5.4", "flask>=3.1.0", "mlflow>=2.19.0", "cloudpickle>=3.1.0", "pandas>=2.2.3", "pyarrow==12.0.0", "pillow==9.1.0", "python-dotenv>=1.0.1"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
